Add and involve Callbackwrapper
The tcpip callback function creates an object instance if the function is a class method. To achieve full object orientated function controll the callback function has to be part of a wrapper class which is cleaned up by the main class destructor to remove the obj instance from the tcpip object.